requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-taggit==1.3.0
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.2.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
